docs(engine): document the Rust surface and make both lint gates run
The Rust crates carried 179 undocumented public items and `missing_docs` was
declared in none of them. `docs-strict.yml` builds `cargo doc --no-deps` for all
four crates and requires it to pass, which it did: `cargo doc` emits a page for
an undocumented item without complaint. The workflow was enforcing that the
reference builds, and that was being read as evidence that the documentation
exists.

All 179 are now written — 70 in the engine, 66 in `program_ad_replay`, 35 in
`studio_wasm_kernel`, 8 in `studio_program_ad_wasm` — and each crate declares
`#![deny(missing_docs)]`, so the closure is permanent instead of a number to
re-check. Every FFI status code documents the condition that produces it, traced
from its return sites rather than restated from its name. Part of this surface
is a `pyo3` extension, so it is public API reached from Python.

Each of the four denials was verified by removing one doc comment and watching
the build fail, then restoring it. A gate never seen to fail is not evidence.

Clippy was not run anywhere: not in CI, not in pre-commit, and no `[lints]`
section existed. It had 26 findings across all targets. Fifteen were real and
are fixed: five complex PyO3 return types now sit behind named aliases following
the crate's existing `analog.rs` convention, a stage-row copy uses
`split_at_mut` instead of an index loop, `as_chunks::<2>` states the pairing in
the type, plus a redundant clone, an `i32`-to-`i32` cast, three heap vectors
that are fixed-size fixtures, and two index loops.

The other eleven are deliberate and were not "fixed". They are row-major indices
written out as `row * dim + col` so each assertion names the element it checks;
collapsing `0 * dim` would hide exactly what the test is about. They carry a
scoped `#[expect]` with that reason — `expect`, not `allow`, so it also warns if
the exception ever stops being needed.

`[lints.clippy] all = deny` now lives in each crate, and a `rust-lint` job runs
clippy and `cargo fmt --check` over all four. The declaration without the job
would have repeated the very failure above: a rule nothing executes. The job
sits inside the reusable static-analysis workflow that every other job needs, so
a finding blocks the pipeline rather than colouring one leg. It is registered in
`ci_workflow_policy.json` and the inventory guard moved 190 -> 191.

Verified: clippy 0 across all targets from the manifests alone, with no `-D`
flag; `cargo fmt --check` clean; 260 engine tests, 15 in `studio_wasm_kernel` and
5 in `studio_program_ad_wasm` pass; the 18 workflow-modularity tests pass.

Parity: this is the Rust half of the documentation rule applied to Python in
`14f624ff`. TypeScript and Julia are measured but not yet gated; their figures
are heuristic and a ratchet set on an estimate would enforce the estimate.
Benchmark: N/A, no benchmarked path is touched.

Seat: 2357035

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ scpn_quantum_engine/src/sectors.rs
@@ -174,7 +174,7 @@ mod tests {
         let n = 2;
         let dim = 1usize << n;
         let mut labels = Vec::with_capacity(dim);
-        let n_i32 = n as i32;
+        let n_i32 = n;
         for k in 0..dim {
             let popcount = (k as u64).count_ones() as i32;
             labels.push(n_i32 - 2 * popcount);
